backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.3.1
django-phonenumbers==1.0.1
django-ratelimit==4.1.0
django-redis==5.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
flower==2.0.1
frozenlist==1.7.0
geoip2==5.1.0
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
kombu==5.5.4
maxminddb==2.8.2
multidict==6.6.3
packaging==25.0
phonenumbers==9.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.2.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1